+++ Suni/Configuration/DatabaseConfiguration.cs
@@ -1,0 +1,12 @@
+using Suni.Suni.Functions.DB;
+
+namespace Suni.Suni.Configuration;
+
+public class DatabaseConfiguration
+{
+    public static void Configure()
+    {
+        var db = new DBMethods();
+        db.Setup();
+    }
+}
