+++ Suni/Functions/Db/DbFilePath.cs
@@ -1,5 +1,6 @@
-namespace Sun.Functions.DB;
+namespace Suni.Suni.Functions.DB;
 
-public partial class DBMethods{
+public partial class DBMethods
+{
     public string dbFilePath = "./Suni/Functions/DB/database.db";
 }
