+++ Suni/Functions/Db/GetUserPrimaryLangAsync.cs
@@ -1,5 +1,5 @@
 using System.Data.SQLite;
-namespace Sun.Functions.DB;
+namespace Suni.Suni.Functions.DB;
 
 public partial class DBMethods
 {
