+++ Suni/Functions/Morse.cs
@@ -1,21 +1,21 @@
 using System.Collections.Generic;
 using System.Text.RegularExpressions;
 
-namespace Sun.Functions
+namespace Suni.Suni.Functions
 {
     public partial class Functions
     {
         private static readonly Dictionary<string, char> morseCodeDictionary = new Dictionary<string, char>()
         {
-            { ".-", 'A' }, { "-...", 'B' }, { "-.-.", 'C' }, { "-..", 'D' }, { ".", 'E' }, 
-            { "..-.", 'F' }, { "--.", 'G' }, { "....", 'H' }, { "..", 'I' }, { ".---", 'J' }, 
-            { "-.-", 'K' }, { ".-..", 'L' }, { "--", 'M' }, { "-.", 'N' }, { "---", 'O' }, 
-            { ".--.", 'P' }, { "--.-", 'Q' }, { ".-.", 'R' }, { "...", 'S' }, { "-", 'T' }, 
-            { "..-", 'U' }, { "...-", 'V' }, { ".--", 'W' }, { "-..-", 'X' }, { "-.--", 'Y' }, 
-            { "--..", 'Z' }, { "-----", '0' }, { ".----", '1' }, { "..---", '2' }, { "...--", '3' }, 
-            { "....-", '4' }, { ".....", '5' }, { "-....", '6' }, { "--...", '7' }, { "---..", '8' }, 
-            { "----.", '9' }, { ".-.-.-", '.' }, { "--..--", ',' }, { "..--..", '?' }, { "-.-.--", '!' }, 
-            { "-.--.", '(' }, { "-.--.-", ')' }, { ".----.", '\'' }, { "-..-.", '/' }, { "-.-.-.", ';' }, 
+            { ".-", 'A' }, { "-...", 'B' }, { "-.-.", 'C' }, { "-..", 'D' }, { ".", 'E' },
+            { "..-.", 'F' }, { "--.", 'G' }, { "....", 'H' }, { "..", 'I' }, { ".---", 'J' },
+            { "-.-", 'K' }, { ".-..", 'L' }, { "--", 'M' }, { "-.", 'N' }, { "---", 'O' },
+            { ".--.", 'P' }, { "--.-", 'Q' }, { ".-.", 'R' }, { "...", 'S' }, { "-", 'T' },
+            { "..-", 'U' }, { "...-", 'V' }, { ".--", 'W' }, { "-..-", 'X' }, { "-.--", 'Y' },
+            { "--..", 'Z' }, { "-----", '0' }, { ".----", '1' }, { "..---", '2' }, { "...--", '3' },
+            { "....-", '4' }, { ".....", '5' }, { "-....", '6' }, { "--...", '7' }, { "---..", '8' },
+            { "----.", '9' }, { ".-.-.-", '.' }, { "--..--", ',' }, { "..--..", '?' }, { "-.-.--", '!' },
+            { "-.--.", '(' }, { "-.--.-", ')' }, { ".----.", '\'' }, { "-..-.", '/' }, { "-.-.-.", ';' },
             { "-...-", '=' }, { ".-.-.", '+' }, { "-....-", '-' }, { ".-..-.", '"' }, { ".--.-.", '@' }
         };
