+++ Suni/Functions/Visual/SymmBuilder.cs
@@ -3,7 +3,7 @@ using SixLabors.ImageSharp;
 using SixLabors.ImageSharp.PixelFormats;
 using SixLabors.ImageSharp.Processing;
 
-namespace Sun.Functions.Visual
+namespace Suni.Suni.Functions.Visual
 {
     public partial class CreateImage
     {
@@ -23,47 +23,47 @@ namespace Sun.Functions.Visual
 
                     //flip the left half of the image horizontallyESQUERDA
                     leftHalf.Mutate(ctx => ctx.Flip(FlipMode.Horizontal));
-            
+
                     //define a rectangle for the right half of the image
                     var rightRect = new Rectangle(width / 2, 0, width / 2, height);
 
                     //copy the flipped left half to the right side of the result image
                     resultImage.Mutate(ctx => ctx.DrawImage(leftHalf, rightRect.Location, 1f));
                     break;
-                
+
 
                 case 2://"right":ojjo
                     var rightHalf = image.Clone(ctx => ctx.Crop(new Rectangle(width / 2, 0, width / 2, height)));
                     resultImage.Mutate(ctx => ctx.DrawImage(rightHalf, new Point(width / 2, 0), 1f));
-            
+
                     rightHalf.Mutate(ctx => ctx.Flip(FlipMode.Horizontal));
-            
+
                     var leftRect = new Rectangle(0, 0, width / 2, height);
-            
+
                     resultImage.Mutate(ctx => ctx.DrawImage(rightHalf, leftRect.Location, 1f));
                     break;
-                
+
 
                 case 3://"jojo":
                     var upHalf = image.Clone(ctx => ctx.Crop(new Rectangle(width / 2, 0, width / 2, height)));
                     resultImage.Mutate(ctx => ctx.DrawImage(upHalf, new Point(width / 2, 0), 1f));
-            
+
                     upHalf.Mutate(ctx => ctx.Flip(FlipMode.Vertical));
-            
+
                     var upRect = new Rectangle(0, 0, width / 2, height);
-            
+
                     resultImage.Mutate(ctx => ctx.DrawImage(upHalf, upRect.Location, 1f));
                     break;
-                
+
 
                 case 4://"ojoj":
                     var downHalf = image.Clone(ctx => ctx.Crop(new Rectangle(width / 2, 0, width / 2, height)));
                     resultImage.Mutate(ctx => ctx.DrawImage(downHalf, new Point(width / 2, 0), 1f));
-            
+
                     downHalf.Mutate(ctx => ctx.Flip(FlipMode.Vertical));
-            
+
                     var downRect = new Rectangle(0, 0, width / 2, height);
-            
+
                     resultImage.Mutate(ctx => ctx.DrawImage(downHalf, downRect.Location, 1f));
                     break;
                 default:
